Adding reception filters

+++ examples/ESP32CANAcceptOnlyStandardFilterDemo/ESP32CANAcceptOnlyStandardFilterDemo.ino
@@ -39,7 +39,7 @@ void setup () {
   ACAN_ESP32_Settings settings (DESIRED_BIT_RATE) ;
   settings.mRequestedCANMode = ACAN_ESP32_Settings::LoopBackMode ;
   const ACAN_ESP32_Filter filter = ACAN_ESP32_Filter::acceptStandardFrames () ;
-  const uint16_t errorCode = ACAN_ESP32::can.begin (settings, filter);
+  const uint32_t errorCode = ACAN_ESP32::can.begin (settings, filter);
   if (errorCode == 0) {
     Serial.print("Bit Rate prescaler: ");
     Serial.println(settings.mBitRatePrescaler);
